Fix boolean attributes and permissions
This solves a few problems with the boolean attributes and the key
permission flags. `CKA_EXTRACTABLE` was never set to true, so private
key attributes could never be retrieved. This led to some other problems
found which are now fixed:

* Boolean flags were stored in `WP11_Object.flags` but read from
  `WP11_Object.opFlag`, so they were never readable.
* Default boolean flags were never set for keys.
* `CKA_SENSITIVE` could change from true to false.
* Wrong boolean masks were used when setting some of the attributes.
* `test_attribute_get` would pass when there is a failure.
* Unwrap attribute setter got mixed with wrap.

Note: The `WP11_Object` type will keep the `flags` variable but it is
now unused. It needs to stay because it may have been stored and could
be loaded in from somewhere. It is now for future use.