Login improvements
This has the following changes:

* Adds flags handling for token
* Adds token flags for marking whether the user or SO pin are set
* Add empty pin check
* Load token objects at open instead of login for no pin (because login
  is not required)
* Allow PINs of zero length
* Remove NSS specific CKF_LOGIN_REQUIRED hack
* Remove CKF_LOGIN_REQUIRED if pin is set and empty

The token flags are so that an empty pin can be used. The session count
switches to 1 for NSS to work around a bug at login time. NSS would try
to do an SO login whilst having another RO session open. Which is not
allowed. When max and min session counts are equal to 1, NSS only opens
a single session and forces that into RW mode.